# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample App API swagger documentation",
        "description": "This is a sample API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/role/{roleId}/permission/get": {
            "post": {
                "tags": [
                    "permission"
                ],
                "summary": "Gets role permissions",
                "operationId": "PermissionParentControllerRolePermissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/role/{roleId}/permission/update": {
            "post": {
                "tags": [
                    "permission"
                ],
                "summary": "Updates role permissions",
                "operationId": "PermissionParentControllerUpdateRolePermissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Role's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permissions",
                        "in": "query",
                        "description": "Role permissions",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}